refactor/テストコードのモジュール化によるリファクタリング > user_usecaseのテスト
